delted silly class and moved methods to vision
